Update header style
