+++ src/app/layout/header-nav/header-nav.component.css
@@ -25,9 +25,9 @@ ul{
 #headerNav{
   position: relative;
   width: 100%;
-  height: 80px;
+  height: 60px;
   background-color: #003c6c;
-  line-height: 50px;
+  line-height: 60px;
   transition: 0.8s;
 }
