+++ src/app/layout/header-top/header-top.component.css
@@ -1,13 +1,16 @@
 /*
-header-top.component.css
+* header-top.component.css
+*
+* 1/5 modified by Holly Hao
+* Adjust the height of the headerTop and headerNav
+* Add margin to the headerTop and headerNav
 
-1/5 modified by Holly Hao
-Adjust the height of the headerTop and headerNav
-Add margin to the headerTop and headerNav
-
-1/2 modified by Holly Hao
-Put the hl to the middle place and adjust the place
-
+* 1/2 modified by Holly Hao
+* Put the hl to the middle place and adjust the place
+*
+* author: Yiyun Zheng
+* Date Revised: 2021/1/24
+* Change header style
 */
 
 ul{
@@ -23,6 +26,8 @@ ul{
     width: 100%;
     /*height: 20%;
     line-height: 18%;*/
+    height: 35px;
+    line-height: 5px;
     margin-top: 1%;
     background-color: #ffffff;
     color: #003c6c;
