+++ src/app/layout/header-top/header-top.component.html
@@ -24,11 +24,11 @@
     </ul>
     <ul class="hr">
         <li>
-            <a href="#" (click)='setLanguage("EN")'>English</a>
+            <a (click)='setLanguage("EN")'>English</a>
         </li>
         <li><span>|</span></li>
         <li>
-            <a href="#" (click)='setLanguage("CN")'>简体中文</a>
+            <a (click)='setLanguage("CN")'>简体中文</a>
         </li>
       </ul>
 </div>
