+++ src/app/layout/header/header.component.css
@@ -13,11 +13,11 @@ Adjust z-index
     position: absolute;
     background-color: white;
     display: block;
-    top:0%;
+    top:10%;
     left: 50%;
     transform: translate(-50%, 0);
-    width: 198px;
-    height: 198px;
+    width: 70px;
+    height: 70px;
     /* border: solid 1px red; */
     transition: 0.8s;
     z-index: 3;
@@ -30,7 +30,7 @@ Adjust z-index
     border:solid, 2px black;
     background-color: white;
     width: 100%;
-    height: 150px;
+    height: 110px;
     transition: 1s;
     z-index: 2;
 }
