+++ src/app/layout/header/header.component.html
@@ -6,7 +6,7 @@
 <div id="header-container">
 
     <a href="https://alumni.ucsc.edu/stay-connected/alumni-affinity-groups/chinese-alumni-association.html" >
-        <img id="logo" src="/assets/image/CN_logo.png" alt="CAAlogo">
+        <img id="logo" src="/assets/image/small_logo.png" alt="CAAlogo">
     </a>
     <app-header-top [language]='language' (setLang)="setLanguage($event)" (setPage)="setPagetype($event)"></app-header-top>
     <app-header-nav [language]='language' (setPage)="setPagetype($event)"></app-header-nav>
