+++ src/app/layout/header/header.component.ts
@@ -1,5 +1,5 @@
-import { viewClassName } from '@angular/compiler';
-import { Component, OnInit, HostListener, Input,ViewChild, EventEmitter, Output} from '@angular/core';
+import { Component, OnInit, Input,EventEmitter, Output} from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
 
 /*
 * author: Holly Hao
@@ -10,6 +10,10 @@ import { Component, OnInit, HostListener, Input,ViewChild, EventEmitter, Output}
 * author : Peter Cai
 * date : 2020/11/8
 * Create headerTop and headerNav
+*
+* author: Yiyun Zheng
+* Date Revised: 2021/1/24
+* Move Header change function to main
 */
 
 
@@ -22,6 +26,7 @@ export class HeaderComponent implements OnInit {
   @Input() language: string;
   @Output() setLang = new EventEmitter<boolean>();
   @Output() setPage = new EventEmitter<boolean>();
+  pageType:String;
   
   setLanguage(event) {
     this.setLang.emit(event);
@@ -32,46 +37,12 @@ export class HeaderComponent implements OnInit {
   }
 
 
-  constructor() { }
+  constructor(
+  ) { }
 
   ngOnInit(): void {
   }
-  /*
-  * changeHeader()
-  * Handle header animation.
-  */
-  @HostListener("document:scroll") changeHeader(){
-    let headerContainer = document.getElementById("header-container");
-    let logoContainer = document.getElementById("logo") as HTMLImageElement;
-    let headerNav = document.getElementById("headerNav");
-    let headerTop = document.getElementById("headerTop");
-    if(document.body.scrollTop>0||document.documentElement.scrollTop>0){
-      headerContainer.style.height = "110px";
-      logoContainer.src = "/assets/image/small_logo.png";
-      logoContainer.style.top ="10%";
-      logoContainer.style.height = "70px";
-      logoContainer.style.width = "70px";
 
-      headerNav.style.height = "60px";
-      headerNav.style.lineHeight = "60px";
-
-      headerTop.style.height = "35px";
-      headerTop.style.lineHeight = "5px";
-    } else {
-      headerContainer.style.height = "198px";
-      logoContainer.style.top ="0%";
-      logoContainer.src = "/assets/image/CN_logo.png";
-      logoContainer.style.height = "198px";
-      logoContainer.style.width = "198px";
-
-      headerNav.style.height = "90px";
-      headerNav.style.lineHeight = "90px";
-
-      headerTop.style.height = "60px";
-      headerTop.style.lineHeight = "20px"
-    }
-    
-  }
 
   public flag:number = 0
